requirements.txt
numpy==1.18.5
opencv-python==4.4.0.42
Pillow==7.2.0
pygame==2.0.0
PyOpenGL==3.1.5
